Bulma theme: actually remove cryptic indicator
Should we add a caption explaining the colors somewhere?
Are the colors along with the titles accessible enough?

The value of the "Priority" and "Status" columns is not
really in the rows of the columns, but in the fact that you
can sort the whole column and get priorized reports first,
or reports with a higher-level status.

Still to be enhanced, for sure.